Updated the Message->containsUnicode() method to detect characters which are outside the GSM default character set.
Nexmo claims to be able to support the GSM 03.38 Basic Character Set <http://unicode.org/Public/MAPPINGS/ETSI/GSM0338.TXT> in text (not unicode) messages. Here, I've changed this method to test specifically for these. The old method fails because many of these characters are outside the normal ASCII range, and because the characters cannot be converted safely using ord() and str_split().